chore: refactor base entity
