+++ apps/backend/src/entities/participant.entity.ts
@@ -1,5 +1,5 @@
 import { Collection, Entity, OneToMany, Property } from "@mikro-orm/core";
-import { BaseEntity } from "../base.entity";
+import { BaseEntity } from "../common/entities/base.entity";
 import { Questionnaire } from "./questionnaire.entity";
 import { Carer } from "./carer.entity";
 import { Language } from "./language.entity";
